pyproject.toml
[project]
name = "gimmes"
version = "0.1.0"
description = "Autonomous Claude Code agent system for trading Kalshi prediction markets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28",
    "pydantic>=2.0",
    "typer>=0.15",
    "cryptography>=44.0",
    "aiosqlite>=0.21",
    "python-dotenv>=1.1",
    "websockets>=15.0",
    "rich>=14.0",
    "tomli>=2.2; python_version < '3.11'",
]

[project.scripts]
gimmes = "gimmes.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
    "mypy>=1.15",
    "respx>=0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gimmes"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring Kalshi demo API access",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
